+++ src/meta/MetaUtility.h
@@ -81,7 +81,7 @@ namespace Meta
     AMAROK_EXPORT QString msToPrettyTime( int ms );
     AMAROK_EXPORT QString secToPrettyTime( int seconds );
 
-    AMAROK_EXPORT QString prettyFilesize( int size );
+    AMAROK_EXPORT QString prettyFilesize( quint64 size );
     AMAROK_EXPORT QString prettyBitrate( int bitrate );
 
     AMAROK_EXPORT QString prettyRating( int rating );
